// frontend/lib/api/generated/conferenceRequestStatus.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Asya Backend API
 * API documentation for Asya Backend application
 * OpenAPI spec version: @appVersion@
 */

/**
 * 会议状态
 */
export type ConferenceRequestStatus =
  (typeof ConferenceRequestStatus)[keyof typeof ConferenceRequestStatus];

export const ConferenceRequestStatus = {
  PREPARING: "PREPARING",
  RUNNING: "RUNNING",
  COMPLETED: "COMPLETED",
} as const;